Some imperfects eliminated.
